Refactor uploader into modular structure
