+++ README.md
@@ -1,1 +1,9 @@
 # 420wedddata
+
+## Usage
+
+Run the uploader:
+
+```
+python main.py
+```
